net/mlx5: fix metadata sharing between E-Switch and VM
If metadata sharing between E-Switch and VM (multi GVMI sharing) is
enabled, PMD creates the default metadata copy rule from register A
to register C_1 on Tx.
This rule should be created in group 0, but due to an incorrect
condition, it was created in group 1.
It prevented propagation of metadata from an application.

This patch fixes the condition by adding the missing check for
external rules, so egress group translation applies only to
application rules.

Fixes: a78425ba3793 ("net/mlx5: support flow metadata between E-Switch and VM")

Signed-off-by: Gregory Etelson <getelson@nvidia.com>
Acked-by: Dariusz Sosnowski <dsosnowski@nvidia.com>